Allow to multiple approaches creating a connected object
While holding down Alt-key, the user is allowed to click multiple times to create a
new connected object.
The Object will in that case be new positioned and always be connected to the
original father.
Clearance of fatherId from within text field - because the field has the focus and
here it can be checked if Alt-key is released.